requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9
pydantic==2.7.4
jieba==0.42.1
pdfplumber==0.11.0
python-docx==1.1.2
EbookLib==0.18
beautifulsoup4==4.12.3

